+++ packages/nimble-components/src/dialog/tests/dialog.spec.ts
@@ -242,7 +242,8 @@ describe('Dialog', () => {
         await disconnect();
     });
 
-    it('focuses the first button on the dialog when it opens', async () => {
+    // Firefox skipped, see: https://github.com/ni/nimble/issues/1075
+    it('focuses the first button on the dialog when it opens #SkipFirefox', async () => {
         const { element, connect, disconnect } = await setup();
         await connect();
         const okButton = document.getElementById('ok')!;
@@ -254,7 +255,8 @@ describe('Dialog', () => {
         await disconnect();
     });
 
-    it('focuses the button with autofocus when the dialog opens', async () => {
+    // Firefox skipped, see: https://github.com/ni/nimble/issues/1075
+    it('focuses the button with autofocus when the dialog opens #SkipFirefox', async () => {
         const { element, connect, disconnect } = await setup();
         await connect();
         const cancelButton = document.getElementById('cancel')!;
@@ -268,7 +270,8 @@ describe('Dialog', () => {
         await disconnect();
     });
 
-    it('supports opening multiple dialogs on top of each other', async () => {
+    // Firefox skipped, see: https://github.com/ni/nimble/issues/1075
+    it('supports opening multiple dialogs on top of each other #SkipFirefox', async () => {
         const { element, connect, disconnect } = await setup();
         await connect();
         const secondDialog = document.createElement('nimble-dialog');