+++ packages/nimble-components/src/icon-base/icon-metadata.ts
@@ -430,6 +430,15 @@ export const iconMetadata: {
     IconTrueFalseRectangle: {
         tags: ['tdms-boolean-channel']
     },
+    IconTriangleTwoLinesHorizontal: {
+        tags: ['collapse-all']
+    },
+    IconTwoSquaresInBrackets: {
+        tags: ['group-by']
+    },
+    IconTwoTrianglesBetweenLines: {
+        tags: ['size-column-to-content']
+    },
     IconUnlink: {
         tags: ['link-broken']
     },