+++ packages/nimble-components/src/table/components/cell/styles.ts
@@ -6,7 +6,15 @@ export const styles = css`
     ${display('grid')}
 
     :host {
+        --ni-private-table-cell-nesting-level: 0;
         padding: 0px calc(${standardPadding} / 2);
+        padding-left: calc(
+            ${standardPadding} / 2 +
+                (
+                    var(--ni-private-table-cell-nesting-level) *
+                        ${standardPadding} * 2
+                )
+        );
         align-self: center;
         height: 100%;
         grid-template-columns: 1fr auto;
@@ -14,7 +22,7 @@ export const styles = css`
         --ni-private-table-cell-action-menu-display: block;
     }
 
-    .cell-content-container {
+    .cell-view {
         overflow: hidden;
         display: flex;
         align-items: center;