+++ packages/nimble-components/src/table/tests/table.pageobject.ts
@@ -4,6 +4,9 @@ import type { TableRecord } from '../types';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import type { MenuButton } from '../../menu-button';
 import type { TableCell } from '../components/cell';
+import type { TableGroupHeaderView } from '../../table-column/base/group-header-view';
+import { TableCellView } from '../../table-column/base/cell-view';
+import type { TableRow } from '../components/row';
 
 /**
  * Page object for the `nimble-table` component to provide consistent ways
@@ -20,14 +23,7 @@ export class TablePageObject<T extends TableRecord> {
     }
 
     public getRenderedCellCountForRow(rowIndex: number): number {
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
-        if (rowIndex >= rows.length) {
-            throw new Error(
-                'Attempting to index past the total number of rendered rows'
-            );
-        }
-
-        const row = rows.item(rowIndex);
+        const row = this.getRow(rowIndex);
         const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
         return cells.length;
     }
@@ -77,23 +73,67 @@ export class TablePageObject<T extends TableRecord> {
         ).length;
     }
 
+    public getRenderedGroupRowCount(): number {
+        return this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-group-row'
+        ).length;
+    }
+
+    public getAllGroupRowExpandedState(): boolean[] {
+        const groupRows = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-group-row'
+        );
+        return Array.from(groupRows).map(row => row.expanded);
+    }
+
+    public getRenderedCellView(
+        rowIndex: number,
+        columnIndex: number
+    ): TableCellView {
+        const cell = this.getCell(rowIndex, columnIndex);
+        const cellView = cell.shadowRoot!.firstElementChild;
+        if (!(cellView instanceof TableCellView)) {
+            throw new Error(
+                'Cell view not found in cell - ensure cellViewTag is set for column'
+            );
+        }
+        return cellView as TableCellView;
+    }
+
     public getRenderedCellContent(
         rowIndex: number,
         columnIndex: number
     ): string {
         return (
-            this.getCell(
+            this.getRenderedCellView(
                 rowIndex,
                 columnIndex
             ).shadowRoot!.textContent?.trim() ?? ''
         );
     }
 
-    public getCellTitle(rowIndex: number, columnIndex: number): string {
+    public getRenderedGroupHeaderContent(groupRowIndex: number): string {
         return (
-            this.getCell(rowIndex, columnIndex)
-                .shadowRoot!.querySelector('.cell-content-container span')
-                ?.getAttribute('title') ?? ''
+            this.getGroupRowHeaderView(
+                groupRowIndex
+            ).shadowRoot!.textContent?.trim() ?? ''
+        );
+    }
+
+    public getAllRenderedGroupHeaderContent(): string[] {
+        const groupRows = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-group-row'
+        );
+        return Array.from(groupRows).map((_, i) => {
+            return this.getRenderedGroupHeaderContent(i);
+        });
+    }
+
+    public getCellTitle(rowIndex: number, columnIndex: number): string {
+        const cellView = this.getRenderedCellView(rowIndex, columnIndex);
+        return (
+            cellView.shadowRoot!.querySelector('span')?.getAttribute('title')
+            ?? ''
         );
     }
 
@@ -102,20 +142,31 @@ export class TablePageObject<T extends TableRecord> {
         columnIndex: number,
         event: Event
     ): boolean | undefined {
-        return this.getCell(rowIndex, columnIndex)
-            .shadowRoot!.querySelector('.cell-content-container span')
+        const cellView = this.getRenderedCellView(rowIndex, columnIndex);
+        return cellView.shadowRoot!.querySelector('span')?.dispatchEvent(event);
+    }
+
+    public getGroupHeaderTitle(groupRowIndex: number): string {
+        const groupHeader = this.getGroupRowHeaderView(groupRowIndex);
+        return (
+            groupHeader
+                .shadowRoot!.querySelector('span')
+                ?.getAttribute('title') ?? ''
+        );
+    }
+
+    public dispatchEventToGroupHeader(
+        groupRowIndex: number,
+        event: Event
+    ): boolean | undefined {
+        const groupHeader = this.getGroupRowHeaderView(groupRowIndex);
+        return groupHeader
+            .shadowRoot!.querySelector('span')
             ?.dispatchEvent(event);
     }
 
     public getRecordId(rowIndex: number): string | undefined {
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
-        if (rowIndex >= rows.length) {
-            throw new Error(
-                'Attempting to index past the total number of rendered rows'
-            );
-        }
-
-        return rows.item(rowIndex).recordId;
+        return this.getRow(rowIndex).recordId;
     }
 
     public getRowWidth(): number {
@@ -132,13 +183,7 @@ export class TablePageObject<T extends TableRecord> {
             );
         }
 
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
-        if (rowIndex >= rows.length) {
-            throw new Error(
-                'Attempting to index past the total number of rendered rows'
-            );
-        }
-        const row = rows[rowIndex];
+        const row = this.getRow(rowIndex);
         const cells = row?.shadowRoot?.querySelectorAll('nimble-table-cell');
         if (columnIndex >= (cells?.length ?? 0)) {
             throw new Error(
@@ -194,16 +239,8 @@ export class TablePageObject<T extends TableRecord> {
     }
 
     public setRowHoverState(rowIndex: number, hover: boolean): void {
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
-        if (rowIndex >= rows.length) {
-            throw new Error(
-                'Attempting to index past the total number of rendered rows'
-            );
-        }
-
-        const cells = rows
-            .item(rowIndex)
-            .shadowRoot!.querySelectorAll('nimble-table-cell');
+        const row = this.getRow(rowIndex);
+        const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
         if (hover) {
             cells.forEach(cell => cell.style.setProperty(
                 '--ni-private-table-cell-action-menu-display',
@@ -216,7 +253,36 @@ export class TablePageObject<T extends TableRecord> {
         }
     }
 
-    private getCell(rowIndex: number, columnIndex: number): TableCell {
+    public async clickRow(rowIndex: number): Promise<void> {
+        const row = this.getRow(rowIndex);
+        row.click();
+        await waitForUpdatesAsync();
+    }
+
+    public getIsRowSelectable(rowIndex: number): boolean {
+        const row = this.getRow(rowIndex);
+        return row.selectable;
+    }
+
+    public getIsRowSelected(rowIndex: number): boolean {
+        const row = this.getRow(rowIndex);
+        return row.selected;
+    }
+
+    public toggleGroupRowExpandedState(groupRowIndex: number): void {
+        const groupRows = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-group-row'
+        );
+        if (groupRowIndex >= groupRows.length) {
+            throw new Error(
+                'Attempting to index past the total number of group rows'
+            );
+        }
+
+        groupRows[groupRowIndex]!.click();
+    }
+
+    private getRow(rowIndex: number): TableRow {
         const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
         if (rowIndex >= rows.length) {
             throw new Error(
@@ -224,7 +290,11 @@ export class TablePageObject<T extends TableRecord> {
             );
         }
 
-        const row = rows.item(rowIndex);
+        return rows.item(rowIndex);
+    }
+
+    private getCell(rowIndex: number, columnIndex: number): TableCell {
+        const row = this.getRow(rowIndex);
         const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
         if (columnIndex >= cells.length) {
             throw new Error(
@@ -233,6 +303,20 @@ export class TablePageObject<T extends TableRecord> {
         }
 
         return cells.item(columnIndex);
+    }
+
+    private getGroupRowHeaderView(groupRowIndex: number): TableGroupHeaderView {
+        const groupRows = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-group-row'
+        );
+        if (groupRowIndex >= groupRows.length) {
+            throw new Error(
+                'Attempting to index past the total number of rendered rows'
+            );
+        }
+
+        const groupRow = groupRows[groupRowIndex];
+        return groupRow!.shadowRoot!.querySelector('.group-header-view')!;
     }
 
     private getHeaderContentElement(